+++ src/popover/index.ts
@@ -1,4 +1,4 @@
+export * from "./inline";
 export * from "./popover";
-export * from "./popover-inline";
 export * from "./popover-trigger";
 export * from "./types";
