+++ src/popover/inline/index.ts
@@ -1,1 +1,2 @@
 export * from "./popover-inline";
+export * from "./types";
