+++ stories/popover/popover-inline/popover-inline.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta } from "@storybook/addon-docs/blocks";
 import * as PopoverInlineStories from "./popover-inline.stories";
-import { PropsTable } from "./props-table";
+import { AutoArgTypesTabs } from "../../storybook-common";
 
 <Meta of={PopoverInlineStories} />
 
@@ -26,4 +26,4 @@ The text and icon inherit the font size of the surrounding text.
 
 ## Component API
 
-<PropsTable />
+<AutoArgTypesTabs of={PopoverInlineStories} />
